requirements.txt
pandas~=2.3.3
numpy~=2.2.6
pillow~=12.0.0
scikit-learn~=1.7.2
scipy~=1.15.3
seaborn~=0.13.2
matplotlib~=3.10.8